refactor(otlp): extract wire-format helpers into dedicated module
Extracts low-level OTLP/gRPC wire helpers from `crates/logfwd-output/src/otlp_sink/encode.rs` into a new `crates/logfwd-output/src/otlp_wire.rs` module. The original file size and churn increases the risk during edits. This addresses part of a larger plan to decouple transport, mapping, and wire encoding.

- Created `crates/logfwd-output/src/otlp_wire.rs` with `encode_key_value_*` and `write_grpc_frame` helpers.
- Updated `crates/logfwd-output/src/otlp_sink/encode.rs` to import from the new module instead of inline definition.
- Preserved existing function signatures and behaviour.
- Added tests matching `write_grpc_frame` inside the new `otlp_wire` module tests.
- Re-exports in `otlp_sink/encode.rs` are `pub(crate)` allowing `otlp_sink` to continue working transparently.

Fixes #2451.
