+++ crates/logfwd-output/src/otlp_sink/encode.rs
@@ -1,6 +1,6 @@
 //! OTLP protobuf encoding: batch-level and row-level encoding logic.
 
-use std::io;
+
 
 use arrow::datatypes::DataType;
 use arrow::record_batch::RecordBatch;
@@ -14,6 +14,7 @@ use logfwd_core::otlp::{
 // These are also re-exported as `pub(super)` so the generated encoder
 // (included via `mod generated { include!(...) }`) can reach them through `super::`.
 pub(super) use logfwd_core::otlp::{encode_fixed32, encode_tag, encode_varint};
+pub(crate) use crate::otlp_wire::{encode_key_value_bool, encode_key_value_bytes, encode_key_value_double, encode_key_value_int, encode_key_value_string, write_grpc_frame};
 
 use super::columns::{
     AttrArray, BatchColumns, ColAttr, format_non_string_attr_value, numeric_timestamp_ns,
@@ -490,58 +491,6 @@ pub(super) fn encode_row_as_log_record(
     encode_fixed64(buf, otlp::LOG_RECORD_OBSERVED_TIME_UNIX_NANO, observed_ns);
 }
 
-/// Encode a KeyValue with string AnyValue as an attribute.
-/// `field_number` is the protobuf field tag of the parent message's `attributes` repeated field
-/// (e.g. `LOG_RECORD_ATTRIBUTES` for LogRecord, `RESOURCE_ATTRIBUTES` for Resource).
-/// KeyValue: { key (string), value (AnyValue { string_value }) }
-pub(super) fn encode_key_value_string(
-    buf: &mut Vec<u8>,
-    field_number: u32,
-    key: &[u8],
-    value: &[u8],
-) {
-    let anyvalue_inner = bytes_field_size(otlp::ANY_VALUE_STRING_VALUE, value.len());
-    let kv_inner = bytes_field_size(otlp::KEY_VALUE_KEY, key.len())
-        + bytes_field_size(otlp::KEY_VALUE_VALUE, anyvalue_inner);
-    encode_tag(buf, field_number, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, kv_inner as u64);
-    encode_bytes_field(buf, otlp::KEY_VALUE_KEY, key);
-    encode_tag(buf, otlp::KEY_VALUE_VALUE, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, anyvalue_inner as u64);
-    encode_bytes_field(buf, otlp::ANY_VALUE_STRING_VALUE, value);
-}
-
-/// Encode a KeyValue with bytes AnyValue (`AnyValue.bytes_value`).
-pub(super) fn encode_key_value_bytes(
-    buf: &mut Vec<u8>,
-    field_number: u32,
-    key: &[u8],
-    value: &[u8],
-) {
-    let anyvalue_inner = bytes_field_size(otlp::ANY_VALUE_BYTES_VALUE, value.len());
-    let kv_inner = bytes_field_size(otlp::KEY_VALUE_KEY, key.len())
-        + bytes_field_size(otlp::KEY_VALUE_VALUE, anyvalue_inner);
-    encode_tag(buf, field_number, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, kv_inner as u64);
-    encode_bytes_field(buf, otlp::KEY_VALUE_KEY, key);
-    encode_tag(buf, otlp::KEY_VALUE_VALUE, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, anyvalue_inner as u64);
-    encode_bytes_field(buf, otlp::ANY_VALUE_BYTES_VALUE, value);
-}
-
-/// Encode a KeyValue with int AnyValue (`AnyValue.int_value`).
-pub(super) fn encode_key_value_int(buf: &mut Vec<u8>, field_number: u32, key: &[u8], value: i64) {
-    let anyvalue_inner = 1 + varint_len(value as u64); // tag(1 byte) + varint
-    let kv_inner = bytes_field_size(otlp::KEY_VALUE_KEY, key.len())
-        + bytes_field_size(otlp::KEY_VALUE_VALUE, anyvalue_inner);
-    encode_tag(buf, field_number, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, kv_inner as u64);
-    encode_bytes_field(buf, otlp::KEY_VALUE_KEY, key);
-    encode_tag(buf, otlp::KEY_VALUE_VALUE, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, anyvalue_inner as u64);
-    encode_varint_field(buf, otlp::ANY_VALUE_INT_VALUE, value as u64);
-}
-
 /// Encode one attribute column into `buf` using pre-computed key encoding.
 ///
 /// Avoids recomputing the
@@ -698,59 +647,3 @@ pub(super) fn encode_col_attr(buf: &mut Vec<u8>, field_number: u32, col: &ColAtt
     }
 }
 
-/// Encode a KeyValue with double AnyValue (`AnyValue.double_value`).
-pub(super) fn encode_key_value_double(
-    buf: &mut Vec<u8>,
-    field_number: u32,
-    key: &[u8],
-    value: f64,
-) {
-    let anyvalue_inner = 1 + 8; // tag(1 byte) + fixed64
-    let kv_inner = bytes_field_size(otlp::KEY_VALUE_KEY, key.len())
-        + bytes_field_size(otlp::KEY_VALUE_VALUE, anyvalue_inner);
-    encode_tag(buf, field_number, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, kv_inner as u64);
-    encode_bytes_field(buf, otlp::KEY_VALUE_KEY, key);
-    encode_tag(buf, otlp::KEY_VALUE_VALUE, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, anyvalue_inner as u64);
-    encode_fixed64(buf, otlp::ANY_VALUE_DOUBLE_VALUE, value.to_bits());
-}
-
-/// Encode a KeyValue with boolean AnyValue (`AnyValue.bool_value`).
-pub(super) fn encode_key_value_bool(buf: &mut Vec<u8>, field_number: u32, key: &[u8], value: bool) {
-    let anyvalue_inner = 1 + 1; // tag(1 byte) + varint(1 byte)
-    let kv_inner = bytes_field_size(otlp::KEY_VALUE_KEY, key.len())
-        + bytes_field_size(otlp::KEY_VALUE_VALUE, anyvalue_inner);
-    encode_tag(buf, field_number, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, kv_inner as u64);
-    encode_bytes_field(buf, otlp::KEY_VALUE_KEY, key);
-    encode_tag(buf, otlp::KEY_VALUE_VALUE, otlp::WIRE_TYPE_LEN);
-    encode_varint(buf, anyvalue_inner as u64);
-    encode_varint_field(buf, otlp::ANY_VALUE_BOOL_VALUE, u64::from(value));
-}
-
-/// Write a gRPC length-prefixed message frame into `buf`.
-///
-/// gRPC wire format (per the [gRPC over HTTP/2 specification](https://grpc.io/docs/what-is-grpc/core-concepts/)):
-/// ```text
-/// [1 byte: compressed flag (0 = not compressed, 1 = compressed)]
-/// [4 bytes: big-endian message length]
-/// [N bytes: protobuf message]
-/// ```
-pub(super) fn write_grpc_frame(
-    buf: &mut Vec<u8>,
-    payload: &[u8],
-    compressed: bool,
-) -> io::Result<()> {
-    let len = u32::try_from(payload.len()).map_err(|_e| {
-        io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "gRPC message payload must be < 4 GiB",
-        )
-    })?;
-    buf.clear();
-    buf.push(u8::from(compressed));
-    buf.extend_from_slice(&len.to_be_bytes());
-    buf.extend_from_slice(payload);
-    Ok(())
-}
